perf: 优化全局组织
