+++ src/views/assets/CommandFilter/CommandFilterDetail/Rules.vue
@@ -1,14 +1,14 @@
 <template>
-  <ListTable :table-config="tableConfig" :header-actions="headerActions" />
+  <GenericListTable :table-config="tableConfig" :header-actions="headerActions" />
 </template>
 
 <script>
-import ListTable from '@/components/ListTable'
+import GenericListTable from '@/layout/components/GenericListTable'
 
 export default {
   name: 'Rules',
   components: {
-    ListTable
+    GenericListTable
   },
   props: {
     object: {
