+++ src/views/xpack/GatheredUser/TaskList.vue
@@ -1,13 +1,13 @@
 <template>
-  <ListTable :table-config="tableConfig" :header-actions="headerActions" />
+  <GenericListTable :table-config="tableConfig" :header-actions="headerActions" />
 </template>
 
 <script>
-import ListTable from '@/components/ListTable'
+import GenericListTable from '@/layout/components/GenericListTable'
 
 export default {
   components: {
-    ListTable
+    GenericListTable
   },
   data() {
     const vm = this
